code/requirements.txt
mysql-connector-python
flask
flask-mail
PyPDF2